Ch3Ex15: Adjusts to using double.

+++ StartingOutWithCPP/Ch3Ex15/Ch3Ex15.cpp
@@ -24,7 +24,7 @@ using namespace std;
 // Main
 int main() {
 	// Vars
-	int property;
+	double property;
 
 	// Input
 	cout << "Please enter the property value: ";
